Next

+++ studio-api/components/WebServer/controllers/user/utility.js
@@ -1,6 +1,7 @@
 const debug = require("debug")(
-  "linto:conversation-manager:components:WebServer:controller:user:utility",
+  "linto:components:WebServer:controllers:user:utility",
 )
+const logger = require(`${process.cwd()}/lib/logger/logger`)
 const model = require(`${process.cwd()}/lib/mongodb/models`)
 
 const CONVERSATION_RIGHTS = require(
@@ -33,7 +34,7 @@ async function getUsersListByConversation(userId, conversation, organiaztion) {
     for (const swUser of conversation.sharedWithUsers) {
       let user = await model.users.getById(swUser.userId)
       if (user.length !== 1) {
-        console.log("User not found", swUser.userId)
+        logger.info("User not found", { userId: swUser.userId })
       } else {
         if (
           isShare &&
@@ -53,7 +54,7 @@ async function getUsersListByConversation(userId, conversation, organiaztion) {
     for (const oUser of organiaztion.users) {
       let user = await model.users.getById(oUser.userId)
       if (user.length !== 1) {
-        console.log("User not found", oUser.userId)
+        logger.info("User not found", { userId: oUser.userId })
       } else {
         if (
           isShare &&
@@ -106,44 +107,53 @@ async function removeUserFromPlatform(userId) {
     const conversations = await model.conversations.getByShare(userId)
 
     // Remove the user from the sharedWithUsers array
-    conversations.map(async (conversation) => {
-      conversation.sharedWithUsers = conversation.sharedWithUsers.filter(
-        (user) => user.userId !== userId,
-      )
-
-      // Update the conversation
-      const resultConvoUpdate = await model.conversations.update(conversation)
-      if (resultConvoUpdate.matchedCount === 0) throw new UserError()
-    })
-
-    // Get all organizations the user is part of
-
-    const organizations = await model.organizations.listSelf(userId)
-    organizations.map(async (organization) => {
-      const data = orgaUtility.countAdmin(organization, userId)
-      if (data.adminCount === 1 && data.isAdmin) {
-        const conversations = await model.conversations.getByOrga(
-          organization._id,
-        )
-        conversations.map(async (conversation) => {
-          deleteFile(
-            `${getStorageFolder()}/${conversation.metadata.audio.filepath}`,
-          )
-
-          const resultConvo = await model.conversations.delete(conversation._id)
-          if (resultConvo.deletedCount !== 1) throw new UserError()
-        })
-        // delete orga
-        const resultOrga = await model.organizations.delete(organization._id)
-        if (resultOrga.deletedCount !== 1) throw new UserError()
-      } else if (data.adminCount > 1 || !data.isAdmin) {
-        organization.users = organization.users.filter(
+    await Promise.all(
+      conversations.map(async (conversation) => {
+        conversation.sharedWithUsers = conversation.sharedWithUsers.filter(
           (user) => user.userId !== userId,
         )
-        let resultOperation = await model.organizations.update(organization)
-        if (resultOperation.matchedCount === 0) throw new UserError()
-      }
-    })
+
+        // Update the conversation
+        const resultConvoUpdate = await model.conversations.update(conversation)
+        if (resultConvoUpdate.matchedCount === 0) throw new UserError()
+      }),
+    )
+
+    // Get all organizations the user is part of
+    const organizations = await model.organizations.listSelf(userId)
+    await Promise.all(
+      organizations.map(async (organization) => {
+        const data = orgaUtility.countAdmin(organization, userId)
+        if (data.adminCount === 1 && data.isAdmin) {
+          const conversations = await model.conversations.getByOrga(
+            organization._id,
+          )
+          await Promise.all(
+            conversations.map(async (conversation) => {
+              if (conversation.metadata?.audio?.filepath) {
+                deleteFile(
+                  `${getStorageFolder()}/${conversation.metadata.audio.filepath}`,
+                )
+              }
+
+              const resultConvo = await model.conversations.delete(
+                conversation._id,
+              )
+              if (resultConvo.deletedCount !== 1) throw new UserError()
+            }),
+          )
+          // delete orga
+          const resultOrga = await model.organizations.delete(organization._id)
+          if (resultOrga.deletedCount !== 1) throw new UserError()
+        } else if (data.adminCount > 1 || !data.isAdmin) {
+          organization.users = organization.users.filter(
+            (user) => user.userId !== userId,
+          )
+          let resultOperation = await model.organizations.update(organization)
+          if (resultOperation.matchedCount === 0) throw new UserError()
+        }
+      }),
+    )
 
     return true
   } catch (error) {
